[clang] Refactor: Move CTAD code from SemaTemplate.cpp to a dedicated file, NFC
Split out the Class Template Argument Deduction (CTAD) related code from
SemaTemplate.cpp to a dedicated file.

These code has grown significantly, and moving it to a separate file will
improve code organization.
